Changes from HPC
UV doesn't work as expected on HPC, especially with the torch package.
So, resolving to using conda env on HPC.
Torch will be installed through conda and remaining packages (as of now) are installed through the pyproject toml file using `pip install -e .`.
Documentation still needs update for the change.